[refactor] websocket handler 기능 분리
